safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.11_The_switch_Statement/S12.11_A3_T2-expected.txt

CONSOLE MESSAGE: line 79: SyntaxError: Unexpected token '{'. Expected '(' to start a subject of a 'switch'.
S12.11_A3_T2

PASS Expected parsing failure

TEST COMPLETE
